feat: updated regex for input amounts and delegated expense object creation to use case

+++ app/src/main/java/com/luna/budgetapp/presentation/screen/expensepreset/ExpensePresetRoute.kt
@@ -107,7 +107,7 @@ fun MainContent(
                             if (dialog.selectedPreset == null) {
                                 onEvent(Event.ConfirmDialog(category, type, amount))
                             } else {
-                                onEvent(Event.AddExpense(dialog.selectedPreset, amount.toDouble(), type))
+                                onEvent(Event.AddExpense(dialog.selectedPreset, amount, type))
                             }
                         },
                         isSaving = dialog.isSaving
